requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
openai==1.10.0
anthropic==0.8.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Terminal UI
rich==13.7.0
prompt-toolkit==3.0.43

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# GitHub Integration
PyGithub==2.1.1

# Azure Integration
azure-identity==1.15.0
azure-mgmt-resource==23.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
